\name{Breast}
\Rdversion{1.0}
\alias{Breast}
\docType{data}
\title{Breast cancer cell line data}
\description{ This data set comes from a breast cancer cell line
microarray study. The experiment was done as
follows. First, the MCF-7 breast cancer cell line was treated
with 17 beta-estradiol or ethanol (vehicle control). Then,
samples were harvested at 1, 4, 12, 24, 36 and 48 hours
after treatment. At each time point, M = 8 replicate arrays
were prepared with each array consisting of G = 1901
genes.}

\usage{Breast}

\format{A matrix containing 1901 rows and 50 columns.}

\references{
Lobenhofer, E., Bennett, L., Cable, P., Li, L., Bushel, P., and Afshari, C. (2002), \emph{Regulation of DNA replication fork genes by 17 beta-estradiol.}
Molec. Endocrin., \bold{16}, 1215-1229.

}

\keyword{datasets}






















































