\name{ORIClust-package}
\Rdversion{1.0}
\alias{ORIClust-package}
\docType{package}
\title{
Order-restricted Information Criterion-based Clustering Algorithm
}
\description{
ORIClust is a user-friendly R-based software package
for gene clustering. Clusters are given by genes matched to prespecified
profiles across various ordered treatment groups. It is
particularly useful for analyzing data obtained from short time-course
or dose-response microarray experiments.
}
\details{
\tabular{ll}{
Package: \tab ORIClust\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-05-24\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
The main functions are \code{\link{ORICC1}} and \code{\link{ORICC2}}, see the documentation files with examples.
}
\author{
Tianqing Liu, Nan Lin, Ningzhong Shi and Baoxue Zhang 

Maintainer:  Tianqing Liu <tianqingliu@gmail.com>
}
\references{
Liu, T., Lin, N., Shi, N. and Zhang, B. (2009), Information criterion-based clustering with order-restricted candidate profiles in short time-course microarray experiments, BMC Bioinformatics, 10: 146.
}

\keyword{ package }

